complete splitting-off of the simulation "processor map" into the in-memory adapter
update/rewrite queries used by ui components to inspect state of simulation processors.

adapters attached to simulation store inside makeSimulationStore().

in-memory event persistence adapter's .persistEvent() implemented (simply .deliverEvent()s the event-to-be-persisted to the in-memory event stream process).

inlined all uses of getEventStream() as it should no longer be assumed to exist by default (or even exist as a concept outside of the in-memory adapter's domain).